Add files via upload
The code together in a scriptable file. Plays chess on python chess by Niklas Fiekas. I renamed the chess module as __start and put it into the same IDE directory for ease of implementation. Ai interacts using classes in python chess as it is much faster. Upon running this script, Ai's with currently 9 parameters will be put into bracket matches.